+++ spec/whois/record/parser/responses/whois.arnes.si/status_registered_spec.rb
@@ -50,7 +50,7 @@ describe Whois::Record::Parser::WhoisArnesSi, "status_registered.expected" do
   describe "#expires_on" do
     it do
       expect(subject.expires_on).to be_a(Time)
-      expect(subject.expires_on).to eq(Time.parse("2010-07-19"))
+      expect(subject.expires_on).to eq(Time.parse("2014-07-19"))
     end
   end
   describe "#nameservers" do